refactor(blog): refactor utils to 1 file and refactor.
